script2-3 decorated

+++ parte_3/script_3.py
@@ -1,16 +1,29 @@
 #!/usr/bin/env python3
 import subprocess, os
 import sys
-from pathlib import Path
 
-subprocess.run("sudo apt install -y podman-docker", shell=True)
-subprocess.run("sudo systemctl enable --now podman.socket", shell=True)
-subprocess.run("sudo systemctl start podman.socket", shell=True)
-subprocess.run("echo 'unqualified-search-registries = [\"docker.io\"]' | sudo tee -a /etc/containers/registries.conf", shell=True)
 
-os.chdir("bookinfo/src/reviews")
-pwd = os.getcwd()
-subprocess.run(f"sudo docker run --rm -u root -v {pwd}:/home/gradle/project -w /home/gradle/project gradle:4.8.1 gradle clean build", shell=True)
-os.chdir("../../..")
-subprocess.run("sudo docker compose -f docker-compose.micro.yml build", shell=True)
-subprocess.run("sudo docker compose -f docker-compose.micro.yml up -d", shell=True)
+if len(sys.argv) < 2:
+    print("Usage: python script_2.py [build|run|stop|debug]")
+    sys.exit(1)
+
+cmd = sys.argv[1].lower()
+
+if cmd == "build":
+    subprocess.run("sudo apt install -y podman-docker", shell=True)
+    subprocess.run("echo 'unqualified-search-registries = [\"docker.io\"]' | sudo tee -a /etc/containers/registries.conf", shell=True)
+
+    os.chdir("bookinfo/src/reviews")
+    pwd = os.getcwd()
+    subprocess.run(f"sudo docker run --rm -u root -v {pwd}:/home/gradle/project -w /home/gradle/project gradle:4.8.1 gradle clean build", shell=True)
+    os.chdir("../../..")
+    subprocess.run("sudo docker compose -f docker-compose.micro.yml build", shell=True)
+
+if cmd == "run":
+    subprocess.run("sudo docker compose -f docker-compose.micro.yml up -d", shell=True)
+
+elif cmd == "stop":
+    subprocess.run("sudo docker compose -f docker-compose.micro.yml down", shell=True)
+
+elif cmd == "debug":
+    subprocess.run("sudo docker compose -f docker-compose.micro.yml up", shell=True)
